Replace getModuleDependencies() method to the Downloader, restructure media treatments
