+++ test/unit/renderers/article.renderer.test.ts
@@ -3,7 +3,6 @@ import { jest } from '@jest/globals'
 import { DELETED_ARTICLE_ERROR } from '../../../src/util/const.js'
 import { VisualEditorRenderer } from '../../../src/util/renderers/visual-editor.renderer.js'
 import { setupScrapeClasses } from '../../util.js'
-import { getModuleDependencies } from '../../../src/util/saveArticles.js'
 import { redisStore } from '../bootstrap.js'
 import { RenderOpts } from 'src/util/renderers/abstract.renderer.js'
 
@@ -40,7 +39,7 @@ describe('ArticleRenderer', () => {
     it('should return visualeditor content if the main page flag is true', async () => {
       const { downloader, dump } = await setupScrapeClasses({ format: '' })
       const { data, articleId, articleDetail } = prepareFixtures({ visualeditor: { content: 'Lorem ipsum dolor sit amet' } })
-      const _moduleDependencies = await getModuleDependencies(articleId, downloader)
+      const _moduleDependencies = await downloader.getModuleDependencies(articleId)
       const result = await visualEditorRenderer.render({
         data,
         redisStore,
@@ -62,7 +61,7 @@ describe('ArticleRenderer', () => {
       const { downloader, dump } = await setupScrapeClasses({ format: '' })
       const content = '<body class="mw-body-content">consectetur adipiscing elit</body>'
       const { data, articleId } = prepareFixtures({ visualeditor: { content } })
-      const _moduleDependencies = await getModuleDependencies(articleId, downloader)
+      const _moduleDependencies = await downloader.getModuleDependencies(articleId)
       const result = await visualEditorRenderer.render({
         data,
         redisStore,
@@ -82,7 +81,7 @@ describe('ArticleRenderer', () => {
       const { downloader, dump } = await setupScrapeClasses({ format: '' })
       const htmlBody = '<body>sed do eiusmod tempor incididunt</body>'
       const { data, articleId, articleDetail } = prepareFixtures({ html: { body: htmlBody }, contentmodel: 'wikitext' })
-      const _moduleDependencies = await getModuleDependencies(articleId, downloader)
+      const _moduleDependencies = await downloader.getModuleDependencies(articleId)
       const result = await visualEditorRenderer.render({
         data,
         redisStore,
@@ -104,7 +103,7 @@ describe('ArticleRenderer', () => {
       const { downloader, dump } = await setupScrapeClasses({ format: '' })
       const htmlBody = '<body>ut labore et dolore magna aliqua. Ut enim ad minim veniam</body>'
       const { data, articleId, articleDetail } = prepareFixtures({ html: { body: htmlBody } })
-      const _moduleDependencies = await getModuleDependencies(articleId, downloader)
+      const _moduleDependencies = await downloader.getModuleDependencies(articleId)
       const result = await visualEditorRenderer.render({
         data,
         redisStore,
@@ -125,7 +124,7 @@ describe('ArticleRenderer', () => {
     it('should return empty string if there was an error during article retrievement', async () => {
       const { downloader, dump } = await setupScrapeClasses({ format: '' })
       const { data, articleId, articleDetail } = prepareFixtures({ error: 'Unexpected internal error' })
-      const _moduleDependencies = await getModuleDependencies(articleId, downloader)
+      const _moduleDependencies = await downloader.getModuleDependencies(articleId)
       const result = await visualEditorRenderer.render({
         data,
         redisStore,
